Put everything together.
The major additions here is generating the final SVG in Python, properly
positioning and clipping the temperature map, and putting the county map on
top of it.

One more major addition is using temperature data from the National Weather
Service for nearby out of state cities. This data is less accurate with
regard to time and the exact location of the measuring stations, but it is
accurate enough to serve its purpose of ensuring that every part of the state
has an estimated temperature on the map.